refactor(shipping-assist): move handoff machine APIs
